+++ packages/components/src/Text/Span.story.tsx
@@ -25,11 +25,53 @@
  */
 
 import React from 'react'
-import { renderWithTheme } from '@looker/components-test-utils'
-import { screen } from '@testing-library/react'
-import { Text } from './Text'
+import { Story } from '@storybook/react/types-6-0'
+import { Span, SpanProps } from './Span'
 
-test('Text has default fontSize', () => {
-  renderWithTheme(<Text>Hello</Text>)
-  expect(screen.getByText('Hello')).toHaveStyleRule('font-size: 16px')
-})
+const Template: Story<SpanProps> = (args) => <Span {...args} />
+
+export const Basic = Template.bind({})
+Basic.args = {
+  children: 'Span Text',
+}
+
+export const XXXXLarge = Template.bind({})
+XXXXLarge.args = {
+  ...Basic.args,
+  fontSize: 'xxxxlarge',
+}
+
+export const Bold = Template.bind({})
+Bold.args = {
+  ...Basic.args,
+  fontWeight: 'bold',
+}
+
+export const Variant = Template.bind({})
+Variant.args = {
+  ...Basic.args,
+  variant: 'critical',
+}
+
+export const TextTransform = Template.bind({})
+TextTransform.args = {
+  ...Basic.args,
+  textTransform: 'uppercase',
+}
+
+export const Wrapped = Template.bind({})
+Wrapped.args = {
+  ...Basic.args,
+  breakword: true,
+}
+
+export const TextDecoration = Template.bind({})
+TextDecoration.args = {
+  ...Basic.args,
+  textDecoration: 'line-through',
+}
+
+export default {
+  component: Span,
+  title: 'Span',
+}